requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=2.1.0=pypi_0
antspynet=0.2.3=pypi_0
antspyx=0.4.2=pypi_0
appdirs=1.4.4=pypi_0
asttokens=2.4.1=pypi_0
astunparse=1.6.3=pypi_0
bzip2=1.0.8=h7b6447c_0
ca-certificates=2023.12.12=h06a4308_0
cachetools=5.3.2=pypi_0
certifi=2023.11.17=pypi_0
charset-normalizer=3.3.2=pypi_0
chart-studio=1.1.0=pypi_0
click=8.1.7=pypi_0
cloudpickle=3.0.0=pypi_0
colorama=0.4.6=pypi_0
comm=0.2.1=pypi_0
contourpy=1.2.0=pypi_0
cycler=0.12.1=pypi_0
debugpy=1.8.0=pypi_0
decorator=5.1.1=pypi_0
deprecated=1.2.14=pypi_0
dill=0.3.8=pypi_0
dm-tree=0.1.8=pypi_0
docker-pycreds=0.4.0=pypi_0
dominate=2.9.1=pypi_0
executing=2.0.1=pypi_0
expat=2.5.0=h6a678d5_0
filelock=3.13.1=pypi_0
flatbuffers=23.5.26=pypi_0
fonttools=4.47.2=pypi_0
freesurfer-stats=1.2.1=pypi_0
fsspec=2023.12.2=pypi_0
gast=0.5.4=pypi_0
gitdb=4.0.11=pypi_0
gitpython=3.1.41=pypi_0
glob2=0.7=pypi_0
google-auth=2.27.0=pypi_0
google-auth-oauthlib=1.2.0=pypi_0
google-pasta=0.2.0=pypi_0
grpcio=1.60.1=pypi_0
h5py=3.10.0=pypi_0
humanize=4.9.0=pypi_0
idna=3.6=pypi_0
imageio=2.33.1=pypi_0
ipykernel=6.29.0=pypi_0
ipython=8.20.0=pypi_0
jedi=0.19.1=pypi_0
jinja2=3.1.3=pypi_0
joblib=1.3.2=pypi_0
jsonpatch=1.33=pypi_0
jsonpointer=2.4=pypi_0
jupyter-client=8.6.0=pypi_0
jupyter-core=5.7.1=pypi_0
keras=2.15.0=pypi_0
kiwisolver=1.4.5=pypi_0
lazy-loader=0.3=pypi_0
ld_impl_linux-64=2.38=h1181459_1
libclang=16.0.6=pypi_0
libffi=3.4.4=h6a678d5_0
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libstdcxx-ng=11.2.0=h1234567_1
libuuid=1.41.5=h5eee18b_0
markdown=3.5.2=pypi_0
markdown-it-py=3.0.0=pypi_0
markupsafe=2.1.4=pypi_0
matplotlib=3.8.2=pypi_0
matplotlib-inline=0.1.6=pypi_0
mdurl=0.1.2=pypi_0
ml-dtypes=0.2.0=pypi_0
mpmath=1.3.0=pypi_0
multiprocess=0.70.16=pypi_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.6.0=pypi_0
networkx=3.2.1=pypi_0
nibabel=5.2.0=pypi_0
numpy=1.26.3=pypi_0
nvidia-cublas-cu12=12.1.3.1=pypi_0
nvidia-cuda-cupti-cu12=12.1.105=pypi_0
nvidia-cuda-nvrtc-cu12=12.1.105=pypi_0
nvidia-cuda-runtime-cu12=12.1.105=pypi_0
nvidia-cudnn-cu12=8.9.2.26=pypi_0
nvidia-cufft-cu12=11.0.2.54=pypi_0
nvidia-curand-cu12=10.3.2.106=pypi_0
nvidia-cusolver-cu12=11.4.5.107=pypi_0
nvidia-cusparse-cu12=12.1.0.106=pypi_0
nvidia-nccl-cu12=2.18.1=pypi_0
nvidia-nvjitlink-cu12=12.3.101=pypi_0
nvidia-nvtx-cu12=12.1.105=pypi_0
oauthlib=3.2.2=pypi_0
opencv-python=4.9.0.80=pypi_0
openssl=3.0.12=h7f8727e_0
opt-einsum=3.3.0=pypi_0
packaging=23.2=pypi_0
pandas=2.2.1=pypi_0
parso=0.8.3=pypi_0
patsy=0.5.6=pypi_0
pexpect=4.9.0=pypi_0
pillow=10.2.0=pypi_0
pip=23.3.1=py311h06a4308_0
platformdirs=4.1.0=pypi_0
plotly=5.18.0=pypi_0
prompt-toolkit=3.0.43=pypi_0
protobuf=4.23.4=pypi_0
psutil=5.9.8=pypi_0
ptyprocess=0.7.0=pypi_0
pure-eval=0.2.2=pypi_0
pyasn1=0.5.1=pypi_0
pyasn1-modules=0.3.0=pypi_0
pycompare=1.5.4=pypi_0
pygments=2.17.2=pypi_0
pyparsing=3.1.1=pypi_0
python=3.11.7=h955ad1f_0
python-dateutil=2.8.2=pypi_0
pytorch-msssim=1.0.0=pypi_0
pytz=2023.3.post1=pypi_0
pyyaml=6.0.1=pypi_0
pyzmq=25.1.2=pypi_0
readline=8.2=h5eee18b_0
requests=2.31.0=pypi_0
requests-oauthlib=1.3.1=pypi_0
retrying=1.3.4=pypi_0
rich=13.7.0=pypi_0
rsa=4.9=pypi_0
scikit-image=0.22.0=pypi_0
scikit-learn=1.4.0=pypi_0
scipy=1.12.0=pypi_0
seaborn=0.13.2=pypi_0
sentry-sdk=1.40.0=pypi_0
setproctitle=1.3.3=pypi_0
setuptools=68.2.2=py311h06a4308_0
shellingham=1.5.4=pypi_0
simpleitk=2.3.1=pypi_0
six=1.16.0=pypi_0
smmap=5.0.1=pypi_0
sqlite=3.41.2=h5eee18b_0
stack-data=0.6.3=pypi_0
statannotations=0.6.0=pypi_0
statsmodels=0.14.1=pypi_0
sympy=1.12=pypi_0
tenacity=8.2.3=pypi_0
tensorboard=2.15.1=pypi_0
tensorboard-data-server=0.7.2=pypi_0
tensorflow=2.15.0.post1=pypi_0
tensorflow-estimator=2.15.0=pypi_0
tensorflow-io-gcs-filesystem=0.36.0=pypi_0
tensorflow-probability=0.23.0=pypi_0
termcolor=2.4.0=pypi_0
threadpoolctl=3.2.0=pypi_0
tifffile=2023.12.9=pypi_0
tk=8.6.12=h1ccaba5_0
torch=2.1.2=pypi_0
torchio=0.19.5=pypi_0
torchvision=0.16.2=pypi_0
tornado=6.4=pypi_0
tqdm=4.66.1=pypi_0
traitlets=5.14.1=pypi_0
triton=2.1.0=pypi_0
typer=0.9.0=pypi_0
typing-extensions=4.9.0=pypi_0
tzdata=2023.4=pypi_0
unet=0.7.7=pypi_0
urllib3=2.1.0=pypi_0
visdom=0.2.4=pypi_0
wandb=0.16.2=pypi_0
wcwidth=0.2.13=pypi_0
webcolors=1.13=pypi_0
websocket-client=1.7.0=pypi_0
werkzeug=3.0.1=pypi_0
wheel=0.41.2=py311h06a4308_0
wrapt=1.14.1=pypi_0
xz=5.4.5=h5eee18b_0
zlib=1.2.13=h5eee18b_0
zope-interface=6.1=pypi_0
